Rename file as it's no longer just about OMOP
