+++ cli/tests/conftest.py
@@ -30,7 +30,7 @@ def _omop_files(tmp_path_factory: pytest.TempPathFactory, monkeypatch) -> None:
     :returns ParquetExport: For direct use when the fixture is explicity called.
     """
     tmpdir_extract = tmp_path_factory.mktemp("repo_base")
-    monkeypatch.setattr("core.omop.ParquetExport.root_dir", tmpdir_extract)
+    monkeypatch.setattr("core.exports.ParquetExport.root_dir", tmpdir_extract)
 
 
 @pytest.fixture()
